Removed DetailsFragment
