--- app/src/main/res/layout/fragment_details.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ VITacademics
-  ~ Copyright (C) 2015  Aneesh Neelam <neelam.aneesh@gmail.com>
-  ~ Copyright (C) 2015  Saurabh Joshi <saurabhjoshi94@outlook.com>
-  ~ Copyright (C) 2015  Gaurav Agerwala <gauravagerwala@gmail.com>
-  ~ Copyright (C) 2015  Karthik Balakrishnan <karthikb351@gmail.com>
-  ~ Copyright (C) 2015  Pulkit Juneja <pulkit.16296@gmail.com>
-  ~ Copyright (C) 2015  Hemant Jain <hemanham@gmail.com>
-  ~
-  ~ This file is part of VITacademics.
-  ~ VITacademics is free software: you can redistribute it and/or modify
-  ~ it under the terms of the GNU General Public License as published by
-  ~ the Free Software Foundation, either version 3 of the License, or
-  ~ (at your option) any later version.
-  ~
-  ~ VITacademics is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  ~ GNU General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU General Public License
-  ~ along with VITacademics.  If not, see <http://www.gnu.org/licenses/>.
-  -->
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <android.support.design.widget.TabLayout
-        android:id="@+id/tabs_details"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@color/colorPrimary"
-        android:theme="@style/ThemeOverlay.AppCompat.Dark"
-        app:elevation="4dp"/>
-
-    <android.support.v4.view.ViewPager
-        android:id="@+id/view_pager_details"
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"/>
-</LinearLayout>
